SNOW-3055676 Remove smkId string conversion
