+++ lib/http/base.js
@@ -186,15 +186,6 @@ function parseIfJSONData(response) {
   // SNOW-3157996 content-type can also have an optional parameter, type/subtype;parameter=value
   const isJson = response['headers']['content-type']?.toLowerCase().includes('application/json');
   if (Util.isString(response['data']) && isJson) {
-    // NOTE:
-    // A super hacky solution to patch /queries/v1/query-request response for PUT/GET operations
-    // Returned encryptionMaterial.smkId is an integer, that might be higher than Number.MAX_SAFE_INTEGER
-    // Treating it as integer will result in uploaded files being UNRECOVERABLE
-    //
-    // This should be replaced with new API field that is a string
-    if (response['data'].includes('smkId')) {
-      response['data'] = Util.convertSmkIdToString(response['data']);
-    }
     response['json'] = response['data'];
     response['data'] = JSON.parse(response['data']);
   }
