+++ lib/services/sf.js
@@ -1085,6 +1085,9 @@ StateConnecting.prototype.continue = function () {
   const clientInfo = {
     CLIENT_APP_ID: this.connectionConfig.getClientType(),
     CLIENT_APP_VERSION: this.connectionConfig.getClientVersion(),
+    CLIENT_CAPABILITIES: {
+      SMK_ID_AS_STRING: true,
+    },
     CLIENT_ENVIRONMENT: {
       APPLICATION_PATH: getApplicationPath(),
       CERT_REVOCATION_CHECK_MODE: this.connectionConfig.crlValidatorConfig.checkMode,
